// src/conductor/providers/spotify/types/typed.ts
export declare namespace SpotifyApi {
// ...
  interface PlaylistSnapshotResponse {
    snapshot_id: string;
  }
// ...
  interface ListOfUsersPlaylistsResponse
    extends PagingObject<PlaylistObjectSimplified> {}

  /**
   * Get a list of the current user's playlists
   *
   * GET /v1/me/playlists
   * https://developer.spotify.com/web-api/get-list-users-playlists/
   */
  interface ListOfCurrentUsersPlaylistsResponse
    extends PagingObject<PlaylistObjectSimplified> {}

  /**
   * Get a playlist
   *
   * GET /v1/users/{user_id}/playlists/{playlist_id}
   * https://developer.spotify.com/web-api/get-playlist/
   */
  interface SinglePlaylistResponse extends PlaylistObjectFull {}

  /**
   * Get a playlist's tracks
   *
   * GET /v1/users/{user_id}/playlists/{playlist_id}/tracks
   * https://developer.spotify.com/web-api/get-playlists-tracks/
   */
  interface PlaylistTrackResponse extends PagingObject<PlaylistTrackObject> {}

  /**
   * Create a Playlist
   *
   * POST /v1/users/{user_id}/playlists
   * https://developer.spotify.com/web-api/create-playlist/
   */
  interface CreatePlaylistResponse extends PlaylistObjectFull {}
// ...
  interface PagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string | null;
    offset: number;
    previous: string | null;
    total: number;
  }
// ...
  interface PlaylistObjectFull extends PlaylistBaseObject {
    /**
     * Information about the followers of the playlist.
     */
    followers: FollowersObject;
    /**
     * Information about the tracks of the playlist.
     */
    tracks: PagingObject<PlaylistTrackObject>;
  }

  /**
   * Playlist Object Simplified
   * [](https://developer.spotify.com/web-api/object-model/)
   */
  interface PlaylistObjectSimplified extends PlaylistBaseObject {
    tracks: {
      href: string;
      total: number;
    };
  }
// ...
}